Store input state locally rather than in Redux.
